requirements.txt
flask
google-genai
python-dotenv
boto3
